requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.3.1
asttokens==2.1.0
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
cachetools==4.2.4
certifi==2022.9.24
charset-normalizer==2.1.1
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
exceptiongroup==1.0.4
executing==1.2.0
frozenlist==1.3.3
fsspec==2022.11.0
future==0.18.2
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grpcio==1.51.0
idna==3.4
importlib-metadata==5.0.0
iniconfig==1.1.1
ipykernel==6.17.1
ipython==8.6.0
jedi==0.18.2
jupyter_client==7.4.7
jupyter_core==5.0.0
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.4.1
matplotlib-inline==0.1.6
multidict==6.0.2
nest-asyncio==1.5.6
numpy==1.18.5
oauthlib==3.2.2
packaging==21.3
pandas==1.2.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.4
pluggy==1.0.0
prompt-toolkit==3.0.33
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.1
Pygments==2.13.0
pyparsing==3.0.9
pytest==7.2.0
python-dateutil==2.8.2
pytorch-lightning==1.5.3
pytz==2022.6
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.9.3
seaborn==0.11.1
six==1.16.0
stack-data==0.6.1
tensorboard==2.2.2
tensorboard-plugin-wit==1.8.1
tomli==2.0.1
torch==1.8.1
torchmetrics==0.6.0
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
typing_extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
yarl==1.8.1
zipp==3.10.0
